[language] Native Functions Refactor
This is mostly removing a verification step for native functions that
was not adding anything in terms of safety but may have had some value
in making developers think harder about native functions.
It does not feel the effort, code, maintenence is giving us value.
We should implement at some point signature checks in the proper places
so we have better protection than ever, but for now this is not less
secure than what we have.
The simplification also makes it easier to split the code so that
the declaration of what native functions are accepted now lives inside
the MoveVM.
There is more coming and this is about changes in the
way we do verification/resolution/dispatching.
I think it needs its own PR
