+++ language/move-vm/runtime/src/lib.rs
@@ -17,6 +17,7 @@ mod interpreter;
 mod loader;
 mod move_vm;
 mod native_context;
+mod native_functions;
 mod runtime;
 
 pub use move_vm::*;
